requirements.txt
numpy>=1.18.5
torch==1.10.1
runstats>=1.8.0
pytorch_lightning==1.3.3
h5py==2.10.0
PyYAML>=5.3.1
pyxb==1.2.6
xmltodict
einops==0.3.0 
fastmri==0.1.1
timm==0.4.12 